Add auth token to dataset and connection services

+++ src/app/services/connection.service.ts
@@ -3,12 +3,15 @@ import { ApiService } from './api.service';
 import { map } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 import { Connection, NewConnection } from '../models/connection.model';
+import { AuthService } from './auth.service';
 
 @Injectable()
 export class ConnectionService {
-    constructor(private api: ApiService) {}
+    constructor(private api: ApiService, private auth: AuthService) {}
 
     getAll(): Observable<Connection[]> {
+        if (this.auth.authToken === null) throw Error('User is not logged in');
+
         return this.api
             .get<
                 {
@@ -18,7 +21,7 @@ export class ConnectionService {
                     user: string;
                     password: string;
                 }[]
-            >('connections')
+            >('connections', this.auth.authToken)
             .pipe(
                 map((connections) =>
                     connections.map((connection) => ({
@@ -33,6 +36,8 @@ export class ConnectionService {
     }
 
     get(id: number): Observable<Connection> {
+        if (this.auth.authToken === null) throw Error('User is not logged in');
+
         return this.api
             .get<{
                 id: number;
@@ -40,7 +45,7 @@ export class ConnectionService {
                 host: string;
                 user: string;
                 password: string;
-            }>(`connections/${id}`)
+            }>(`connections/${id}`, this.auth.authToken)
             .pipe(
                 map((connection) => ({
                     id: connection.id,
@@ -53,23 +58,42 @@ export class ConnectionService {
     }
 
     create(connection: NewConnection) {
-        return this.api.post('connections', {
-            connectionName: connection.name,
-            host: connection.host,
-            user: connection.user,
-            password: connection.password,
-        });
+        if (this.auth.authToken === null) throw Error('User is not logged in');
+
+        return this.api.post(
+            'connections',
+            {
+                connectionName: connection.name,
+                host: connection.host,
+                user: connection.user,
+                password: connection.password,
+            },
+            this.auth.authToken
+        );
     }
 
     editName(id: number, newName: string) {
-        return this.api.put(`connections/${id}`, { newName: newName });
+        if (this.auth.authToken === null) throw Error('User is not logged in');
+
+        return this.api.put(
+            `connections/${id}`,
+            { newName: newName },
+            this.auth.authToken
+        );
     }
 
     delete(id: number) {
-        return this.api.delete(`connections/${id}`);
+        if (this.auth.authToken === null) throw Error('User is not logged in');
+
+        return this.api.delete(`connections/${id}`, this.auth.authToken);
     }
 
     getConnectionDatabases(id: number) {
-        return this.api.get<string[]>(`connections/${id}/databaseName`);
+        if (this.auth.authToken === null) throw Error('User is not logged in');
+
+        return this.api.get<string[]>(
+            `connections/${id}/databaseName`,
+            this.auth.authToken
+        );
     }
 }
